Move column visibility feature to a flexlayout-react Tab
