+++ src/Logs/LogLayout.tsx
@@ -6,6 +6,7 @@ import 'flexlayout-react/style/light.css';
 import {Layout1} from '../Layouts/Layout1';
 // split component
 import LogTable from "./LogTable"
+import LogColVisibility from "./LogColVisibility";
 
 export default function LogLayout() {
     // reference for flexlayout (used in the future)
@@ -20,7 +21,7 @@ export default function LogLayout() {
         const component = node.getComponent();
         // render additional components based on value
         switch(component) {
-            case "Placeholder": return <div>{node.getName()}</div>
+            case "ColumnVisibility": return <LogColVisibility/>
             case "Table": return <LogTable/>
         }
     }, [])
